Connect NoPhotoDisclaimer screen
Also added a check for maximum images and a header for NoPhotoDisclaimer with the option to go back to the main screen
