--- src/screens/MainApp/DiagnoseRequest/ImageReview/RenderedImage/index.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { Image } from 'react-native'
-
-const RenderedImage = ({ uri, style }) => {
-  return (
-    <Image
-      resizeMethod='scale'
-      style={style}
-      source={{ uri }}
-    />
-  )
-}
-
-export default RenderedImage
