--- src/screens/MainApp/DiagnoseRequest/ImageReview/constants.js
@@ -1,3 +1,0 @@
-export const MAX_IMAGES = 10
-export const MIN_IMAGES = 3
-export const ADD_MORE_IMAGES_URI = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAIAAAACACAQAAABpN6lAAAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAJcEhZcwAAAEgAAABIAEbJaz4AAAAHdElNRQfjCxIGGiBeCssoAAAAxElEQVR42u3YMU6EUABF0a8xsZ7ObjZg4lZ0GbMB16LbNJNYakNoaWDuBM6hoSDwchMKGAMAAAAAAAAAAAAAAHbnIXz2eXxPZ7/jow5ReB1/03HtRjzWFWoC1ANqAtQDagLUA2oC1ANqAtQDagLUA2pPm935bVwWrjjNZ8/ja/F+n+PnVlHW8T5/7K5zvGwz8/CvgAD1gCPzS+weCFAPqAlQD6gJUA+oCVAPqAlQD6gJUA+oHT4AAAAAAAAAAAAAAAB79A9E6zZrwRW6nAAAACV0RVh0Y3JlYXRlLWRhdGUAMjAxMy0wMS0yMVQyMjo1MDozNS0wNjowMFJsIWUAAAAldEVYdGRhdGU6Y3JlYXRlADIwMTktMTEtMThUMDY6MDQ6MTQrMDA6MDADyak3AAAAJXRFWHRkYXRlOm1vZGlmeQAyMDE5LTEwLTIzVDA1OjIyOjAxKzAwOjAw2h1rmwAAACV0RVh0bW9kaWZ5LWRhdGUAMjAxMy0wMS0yMVQyMjo1MDozNS0wNjowMA3dV1EAAAAASUVORK5CYII='
